upstox_bot/requirements.txt
# Upstox Trading Bot Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
pandas>=2.0.0
numpy>=1.24.0
}